requirements.txt
requests
pandas
python-dotenv
tqdm
fastapi
uvicorn
xgboost
scikit-learn
joblib
pydantic
